requirements.txt
python-docx>=1.1.0
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
